Add clear mention of Phabricator scripts to README
The standard user documentation, if not a terrific read (it is, if you devote some time, like any good book), is a source of extensive knowledge on configuring and managing Phabricator. For a beginner, that succeeds in setting up a Dockerized version of Phabricator (for example to try out the product), it is reasonably hard to follow the docs, that are written from a perspective of a "downloaded directory". While crossing this chasm this can be an entire Wiki (hopefully in the future), one of the most pressing issues in the first-time setup is access to utility scripts Phabricator provides. The README previously had no clear mention of where they are / how to find them even though they are critical to a successful deployment — in fact some values in the config cannot be accessed through the UI *by design.*
- added new section briefly mentioning the location of Phabricator home folder and a link to Phabricator configuration docs